requirements.txt
binwalk==2.1.0
capstone==5.0.0rc2
pyelftools==0.28
keystone-engine==0.9.2
pwntools==4.8.0